rename ResourcesEndpointsE2ESpec, ValuesEndpointsE2ESpec, VersionEndpointE2ESpec, KnoraSipiIntegrationE2ESpec
